+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for croud
 Unreleased
 ==========
 
+- Fixed region support in ``consumers edit`` command.
+
 - Added ``clusters delete`` command to delete existing clusters.
 
 - Added ``clusters scale`` command to scale existing clusters.